requirements.txt
Flask==2.2.5
Flask-MySQLdb==0.2.0
requests==2.26.0
Werkzeug==2.2.3
python-dotenv==1.0.0
